+++ .coveragerc
@@ -5,6 +5,3 @@ exclude_lines =
 
     # Don't complain if tests don't hit defensive assertion code:
     raise NotImplementedError
-
-    # Don't complain about the bodies of abstract functions
-    \.\.\.
